server: move fixture to a specific test file where it's used
